refactor:  Move LimitedDistinctAggregation to physical-optimizer crate
