+++ datafusion/physical-optimizer/src/lib.rs
@@ -19,6 +19,7 @@
 
 pub mod aggregate_statistics;
 pub mod limit_pushdown;
+pub mod limited_distinct_aggregation;
 mod optimizer;
 pub mod output_requirements;
